gmxapi-134 Stop signal handled incorrectly for multiple simulations
Bring sample_restraint repository up-to-date with development in
gromacs-gmxapi and elsewhere.
